chore: reorganise and use dnd-data helpers
